moved shared components out of components
