+++ src/app/app.module.ts
@@ -1,29 +1,14 @@
 // Core angular modules
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { CustomFormsModule } from 'ng2-validation';
-import { HttpModule } from '@angular/http';
 import { AngularFireModule, AuthProviders, AuthMethods } from 'angularfire2';
-import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
 import { RouterModule } from '@angular/router';
 import { StoreLogMonitorModule, useLogMonitor } from '@ngrx/store-log-monitor';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { SlimLoadingBarModule } from 'ng2-slim-loading-bar';
-import { ToastyModule } from 'ng2-toasty';
-import { DatepickerModule } from 'angular2-material-datepicker';
-import { CalendarModule } from 'primeng/primeng';
-
-// Pipes
 
 // Services 
 import { ServiceModule } from './services/index';
-
-// Guards
-import { TripsResolveGuard } from './guards/trips-resolve.guard';
-import { CanActivateViaAuthGuard } from './guards/auth.guard';
-
 import { routes } from './app.routes';
 import { reducer } from './reducers/index';
 
@@ -34,7 +19,6 @@ import { TripHasPlaces } from './Validators/trip-has-places.directive';
 
 // Components
 import { AppComponent } from './app.component';
-import { InstagramAuthenticationCallbackComponent } from './components/shared/instagram-authentication-callback/instagram-authentication-callback.component';
 import { AmbassadorComponent } from './components/misc/ambassador/ambassador.component';
 
 // Satellizer Module
@@ -42,11 +26,10 @@ import { Ng2UiAuthModule, CustomConfig } from 'ng2-ui-auth';
 import { MyAuthConfig } from './auth-config';
 import { AppEffectsModule } from './effects/index';
 import { ComponentsModule } from './components/index';
-import { SharedModule } from './components/shared/index';
-
-
-
-
+import { SharedModule } from './shared/index';
+import { CanActivateViaAuthGuard } from './guards/auth.guard';
+import { TripsResolveGuard } from './guards/trips-resolve.guard';
+import { InstagramAuthenticationCallbackComponent } from './shared/instagram-authentication-callback/instagram-authentication-callback.component';
 const firebaseConfig = {
   apiKey: "AIzaSyDRiL-DZLnvLoj37YZNqQyYcOaOecXFOus",
   authDomain: "travel-app-frontend.firebaseapp.com",
@@ -58,22 +41,17 @@ const myFirebaseAuthConfig = {
   provider: AuthProviders.Google,
   method: AuthMethods.Redirect
 }
-
 @NgModule({
   declarations: [
     AppComponent,
-    InstagramAuthenticationCallbackComponent,
     FileSelectDirective,
     Autosize,
     TripHasPlaces,
     AmbassadorComponent,
+    InstagramAuthenticationCallbackComponent
   ],
   imports: [
     BrowserModule,
-    FormsModule,
-    CustomFormsModule,
-    ReactiveFormsModule,
-    HttpModule,
     Ng2UiAuthModule.forRoot(MyAuthConfig),
     AngularFireModule.initializeApp(firebaseConfig, myFirebaseAuthConfig),
     StoreModule.provideStore(reducer),
@@ -87,18 +65,15 @@ const myFirebaseAuthConfig = {
     //   })
     // }),
     StoreLogMonitorModule,
-    SlimLoadingBarModule.forRoot(),
-    ToastyModule.forRoot(),
-    DatepickerModule,
-    CalendarModule,
+    
     ComponentsModule,
     SharedModule,
     ServiceModule,
     AppEffectsModule
   ],
   providers: [
-    TripsResolveGuard,
-    CanActivateViaAuthGuard
+    CanActivateViaAuthGuard,
+    TripsResolveGuard
   ],
   bootstrap: [AppComponent]
 })
