+++ src/app/components/dashboard/index.ts
@@ -6,7 +6,7 @@ import { DashboardSearchComponent } from './dashboard-search/dashboard-search.co
 import { FeedTripsComponent } from './feed-trips/feed-trips.component';
 import { TrendingTripsComponent } from './trending-trips/trending-trips.component';
 import { DashboardRoutes } from './dashboard.routes';
-import { SharedModule } from '../shared/index';
+import { SharedModule } from '../../shared/index';
 
 import { InfiniteScrollModule } from 'angular2-infinite-scroll';
 import { BrowserModule } from '@angular/platform-browser';
