+++ src/app/components/index.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { SearchModule } from './search/index';
-import { SharedModule } from './shared/index';
 import { UserModule } from './user/index';
 import { DashboardModule } from './dashboard/index';
 import { TripsModule } from './trips/index';
