+++ src/app/components/search/index.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { SearchComponent } from './search.component';
-import { SharedModule } from '../shared/index';
+import { SharedModule } from '../../shared/index';
 import { SearchedTripsComponent } from './searched-trips/searched-trips.component';
 import { TripsSearchComponent } from './trips-search/trips-search.component';
