+++ src/app/shared/header/header.component.scss
