+++ src/app/shared/index.ts
@@ -7,8 +7,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { LoginComponent } from './auth/login/login.component';
 import { SignupComponent } from './auth/signup/signup.component';
 import { HeaderComponent } from './header/header.component';
-import { SearchComponent } from '../search/search.component';
-import { DashboardSearchComponent } from '../dashboard/dashboard-search/dashboard-search.component';
 import { SearchFilterComponent } from './search/search-filter/search-filter.component';
 import { TripsSearchFilterComponent } from './search/trips-search-filter/trips-search-filter.component';
 import { passwordValidator } from './validators/password.validator';
@@ -18,8 +16,11 @@ import { SlimLoadingBarModule } from 'ng2-slim-loading-bar';
 import { LoaderComponent } from './loader/loader.component';
 import { TripListItemComponent } from './trips/trip-list-item/trip-list-item.component';
 import { MomentModule } from 'angular2-moment';
-import { HumanizePipe } from '../../pipes/humanize';
-import { ThumbnailPipe } from '../../pipes/thumbnail';
+import { ToastyModule } from 'ng2-toasty';
+import { DatepickerModule } from 'angular2-material-datepicker';
+import { CalendarModule } from 'primeng/primeng';
+import { HumanizePipe } from '../pipes/humanize';
+import { ThumbnailPipe } from '../pipes/thumbnail';
 
 const routes: Routes = [
 ];
@@ -45,7 +46,11 @@ const routes: Routes = [
     ReactiveFormsModule,
     SlimLoadingBarModule.forRoot(),
     CommonModule,
-    MomentModule
+    MomentModule,
+    SlimLoadingBarModule.forRoot(),
+    ToastyModule.forRoot(),
+    DatepickerModule,
+    CalendarModule
   ],
   exports: [
     // components
@@ -64,7 +69,11 @@ const routes: Routes = [
     
     // modules
     CommonModule,
-    MomentModule
+    MomentModule,
+    ToastyModule,
+    SlimLoadingBarModule,
+    DatepickerModule,
+    CalendarModule,
   ]
 })
 export class SharedModule {}
