+++ src/app/shared/loader/loader.component.scss
