+++ src/app/shared/search/dashboard-search-filter/dashboard-search-filter.component.scss
