+++ src/app/shared/search/trips-search-filter/trips-search-filter.component.scss
