+++ src/app/shared/trips/trip-list-item/trip-list-item.component.html
